requirements.txt
pandas
jupyter
sqlalchemy
pyyaml
psycopg2
kaggle
mysql-connector-python
matplotlib
seaborn
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-pandas
pandas
numpy
requests
grafana-client

